Move SubLocationsDropdown functionality within summary to its own component, save sub locations against the batch when the multi select is blurred
